Add first version of package browsing experience
Adds a new command that lists all azure-sdk-for-node packages from npm
and user can filter the list and select a particular package. Hitting
Enter after selection adds a dependency to the selected package in the
project's package.json and adds a require statement in the active
document, if one is open.